+++ Sources/ToastUI/ToastViewStyle.swift
@@ -9,11 +9,14 @@ import SwiftUI
 
 /// The properties of a `ToastView` instance.
 public struct ToastViewStyleConfiguration {
-  /// Content of the `ToastView`.
-  public let content: AnyView?
+  /// A type-erased view describes the background of `ToastView`.
+  public var background: AnyView?
 
-  /// A label describes the `ToastView`.
-  public let label: Text?
+  /// A type-erased view describes the label of `ToastView`.
+  public var label: AnyView?
+
+  /// A type-erased view describes the content of `ToastView`.
+  public var content: AnyView?
 }
 
 /// A type that applies standard interaction behavior to all `ToastView`s
@@ -40,12 +43,12 @@ extension ToastViewStyle {
   }
 }
 
-/// A type-erased `ToastViewStyle`.
+/// A concrete, type-erased `ToastViewStyle`.
 public struct AnyToastViewStyle: ToastViewStyle {
   private let _makeBody: (Configuration) -> AnyView
 
-  /// Creates a type-erased `ToastViewStyle`.
-  public init<Style: ToastViewStyle>(_ style: Style) {
+  /// Creates a concrete, type-erased `ToastViewStyle`.
+  public init<Style>(_ style: Style) where Style: ToastViewStyle {
     self._makeBody = style.eraseToAnyView
   }